+++ vessel-runtime/src/main/java/com/textnow/android/vessel/VesselImpl.kt
@@ -48,13 +48,15 @@ import kotlin.reflect.KClass
  * @param inMemory true if the database should be in-memory only [Default: false]
  * @param allowMainThread to allow calls to be made on the main thread. [Default: false]
  * @param callback for notifications of database state changes
+ * @param cache Optional [VesselCache]. Built-ins: [DefaultCache] and [LruCache] [Default: null]
  */
 class VesselImpl(
         private val appContext: Context,
         private val name: String = "vessel-db",
         private val inMemory: Boolean = false,
         private val allowMainThread: Boolean = false,
-        private val callback: VesselCallback? = null
+        private val callback: VesselCallback? = null,
+        private val cache: VesselCache? = null
 ): Vessel {
     // region initialization
 
@@ -142,12 +144,11 @@ class VesselImpl(
      */
     override fun <T : Any> getBlocking(type: KClass<T>): T? {
         check(!closeWasCalled) { "Vessel($name:${hashCode()}) was already closed." }
-        type.qualifiedName?.let { typeName ->
-            dao.getBlocking(typeName)?.data?.let { entity ->
-                return fromJson(entity, type)
+        return type.qualifiedName?.let { typeName ->
+            cache?.get(typeName) ?: dao.getBlocking(typeName)?.data?.let { entity ->
+                fromJson(entity, type)
             }
         }
-        return null
     }
 
     /**
@@ -158,12 +159,11 @@ class VesselImpl(
      */
     override fun <T : Any> getBlocking(type: Class<T>): T? {
         check(!closeWasCalled) { "Vessel($name:${hashCode()}) was already closed." }
-        type.kotlin.qualifiedName?.let { typeName ->
-            dao.getBlocking(typeName)?.data?.let { entity ->
-                return fromJson(entity, type.kotlin)
+        return type.kotlin.qualifiedName?.let { typeName ->
+            cache?.get(typeName) ?: dao.getBlocking(typeName)?.data?.let { entity ->
+                fromJson(entity, type.kotlin)
             }
         }
-        return null
     }
 
     /**
@@ -174,6 +174,7 @@ class VesselImpl(
     override fun <T : Any> setBlocking(value: T) {
         check(!closeWasCalled) { "Vessel($name:${hashCode()}) was already closed." }
         typeNameOf(value).let { typeName ->
+            cache?.set(typeName, value)
             dao.setBlocking(entity = VesselEntity(
                     type = typeName,
                     data = toJson(value)
@@ -189,6 +190,7 @@ class VesselImpl(
     override fun <T : Any> deleteBlocking(type: KClass<T>) {
         check(!closeWasCalled) { "Vessel($name:${hashCode()}) was already closed." }
         type.qualifiedName?.let { typeName ->
+            cache?.remove(typeName)
             dao.deleteBlocking(typeName)
         }
     }
@@ -201,6 +203,7 @@ class VesselImpl(
     override fun <T : Any> deleteBlocking(type: Class<T>) {
         check(!closeWasCalled) { "Vessel($name:${hashCode()}) was already closed." }
         type.kotlin.qualifiedName?.let { typeName ->
+            cache?.remove(typeName)
             dao.deleteBlocking(typeName)
         }
     }
@@ -217,12 +220,11 @@ class VesselImpl(
      */
     override suspend fun <T : Any> get(type: KClass<T>): T? {
         check(!closeWasCalled) { "Vessel($name:${hashCode()}) was already closed." }
-        type.qualifiedName?.let { typeName ->
-            dao.get(typeName)?.data?.let { entity ->
-                return fromJson(entity, type)
+        return type.qualifiedName?.let { typeName ->
+            cache?.get(typeName) ?: dao.get(typeName)?.data?.let { entity ->
+                fromJson(entity, type)
             }
         }
-        return null
     }
 
     /**
@@ -233,6 +235,7 @@ class VesselImpl(
     override suspend fun <T : Any> set(value: T) {
         check(!closeWasCalled) { "Vessel($name:${hashCode()}) was already closed." }
         typeNameOf(value).let { typeName ->
+            cache?.set(typeName, value)
             dao.set(entity = VesselEntity(
                 type = typeName,
                 data = toJson(value)
@@ -248,6 +251,7 @@ class VesselImpl(
     override suspend fun <T : Any> delete(type: KClass<T>) {
         check(!closeWasCalled) { "Vessel($name:${hashCode()}) was already closed." }
         type.qualifiedName?.let { typeName ->
+            cache?.remove(typeName)
             dao.delete(typeName)
         }
     }
@@ -284,13 +288,17 @@ class VesselImpl(
         val newName = typeNameOf(new)
         oldType.qualifiedName?.let { oldName ->
             if (oldName == newName) {
+                cache?.set(newName, new)
                 set(new)
             } else {
+                cache?.remove(oldName)
+                cache?.set(newName, new)
                 dao.replace(
                     oldType = oldName,
                     new = VesselEntity(
                         type = newName,
-                        data = toJson(new))
+                        data = toJson(new)
+                    )
                 )
             }
         }
@@ -301,6 +309,7 @@ class VesselImpl(
      */
     override fun clear() {
         check(!closeWasCalled) { "Vessel($name:${hashCode()}) was already closed." }
+        cache?.clear()
         db.clearAllTables()
     }